pyproject.toml
[tool.poetry]
name = "image-api"
version = "0.1.0"
description = ""
authors = ["Blacks7566 <nitinsoni815@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
Pillow = "^10.0.0"
python-multipart = "^0.0.6"
boto3 = "^1.28.35"
pydantic-settings = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
